Fix the size to be cleaned in prepare_inputs
In dudect/constant.c, to do the test, we must first distribute all
100000 data into two classes, fixed and random class. If the one is
distributed to 0, it means it belongs to random class and we should
clean the relating input_data to 0.

But in this Implementation, a single size of an element in input_data is
`chunk_size` which means it is 16 byte. However, in dudect/constant.c,
it only clean up an `uint16_t` size, which is 2 byte only.

To fix this problem, I use memset to reset the whole element to 0.