Apply X-macro trick to reduce duplicated code
X macros are a technique for reliable maintenance of parallel lists
of code and/or data, whose corresponding items must be declared or
executed in the same order.

The operations in DUT are defined and implemented by means of X-macro
technique, meaning that the duplicated code can then be eliminated.

Reference: https://en.wikipedia.org/wiki/X_Macro