Add new command and associated test
This commit supports a new command 'merge', which
merges queues into a queue in ascending order.
Even if queues are not sorted, 'merge' can be called.
However, we will get the wrong answer, which is
detected immediately after the call.

Besides, this commit forwards the declaration
of 'qcontext_t' from qtest.c to queue.h.
It is because that 'q_merge' needs to traverse the chain.

Relative test bench is added into trace-03-ops.cmd