Register `line_atexit` in `linenoise` function
Before this change, `line_atexit` is registered in `enable_raw_mode`.
However, if stdin is not a tty, (eg. it may be a file),
`enable_raw_mode` will not be called and `line_atexit` will not be
registered.

In this case, if we use valgrind to detect memory leak, the history will
not be freed if stdin is not a tty. (eg. using trace files as stdin)

This commit fixes this problem. It make sure that `line_atexit` is
registered when linenoise is used.