Fix pre-push.hook validation bug
We utilize ./scripts/pre-push.hook to do the checking when trying to do
`git push`. One of the stages is to check a speicific commit hash number
for validation of the repository.

It tries to get the hash number of commit
3ed17237af5b1ead6c394df5099bc2bc1f8392df, using a filter to get the
value of `commit` whose git log message matches the pattern specified
 in line 15 of `/scripts/pre-push.hook`.
However when merging the commit into the master branch, it generated a
new commit which is 390ade9eca44b432acb758084e7122cc2c46e485, so the
value of `commit` in line 15 of `/scripts/pre-push.hook` will be
incorrect.

That's why we modify a small part of the command in line 15 from `-n 1`
to `--skip 1`, so we can neglect the merging commit and get the actual
commit hash number we want.

Fix #151