Implement q_new function
According to the materials from lab0 provided by the instructor, it's understood
that the definitions of NULL queue and Empty queue are different. An Empty queue
 signifies that the queue points to a valid data structure (head). Therefore, when
implementing this method, I need to allocate a memory block for head using malloc
and then set the prev and next pointers of head to point to itself.