Group github action log lines
Previously, the workflow used to test the code does not group log
messages, possibly making the output too lengthy and hard to scroll on
the web. Especially the logs generated by dudect/fixture.c, although
it shows perfectly fine on local machines by rewriting previous lines
in the terminal, github actions does not seem to support it.

The log grouping is done by adding "::group::{title}" and
"::endgroup::" which are mentioned in github docs. Another argument
"--group-output" is added to "scripts/driver.py" to optionally enable
grouping. Added another target "test_ga" in "Makefile" and replaced
"make test" with it in github workflow. ("ga" is the abbreviation for
"github actions")

Link: https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/workflow-commands-for-github-actions#grouping-log-lines
Change-Id: I77d4455967b6d7e1d118049d91e67346f6dcdc1a