Introduce EditorConfig
Originally, the indentation of the scripts was not unified: ones in
'scripts/' used 2 spaces, and ones in '.ci/' used 4. Even though the
former used 2 spaces, the style was not implemented well as some indents
in the files were 4 spaces.

This patch unifies the style of the scripts, and note it with an
EditorConfig file. EditorConfig is used because it's supported
out-of-the-box by many editors, helping the auto-indent functionality
works.

Change-Id: I41be8779ef5b0e94192c144bb48605625727963f