Fix BSD compatibility
On macOS, Perl-compatible regex option (-P) in GNU grep is not available
by default. It can be replaced with sed. While using 'tr', it is
necessary to set the locale to "C" ensures that tr interprets the byte
stream in a consistent, ASCII-only manner. This prevents errors like
"Illegal byte sequence" on BSD systems such as macOS.

Change-Id: I1c134092a5fb2afc8018a638e172fb0b380e0fdd