Introduce EditorConfig
Originally, the indentation of the scripts in directory 'scripts/' and
the GitHub workflow was inconsistent due to human error, and some files
contains lines with only spaces.

This patch introduces EditorConfig, which is supported out-of-the-box by
many editors, to help the auto-indent functionality work. It sets the
indentation to the currently used indentation of each file respectively,
and ask editors to use Unix-like newlines (line feed at the end of each
line) and avoid trailing white-spaces.

Change-Id: Idaa878cb779b637c1361baa637419645828530cb