Introduce REST API fallback
Switch to using the GitHub REST API as a fallback when HTML parsing
fails to retrieve the commit hash from the GitHub commits page. This
ensures that the script continues to work even under rate limit
conditions and improves reliability. Optionally, authentication can
be used to further mitigate rate limits.

Co-authored-by: Po-Ying Chiu <charlie910417@gmail.com>
Change-Id: I8688fdb093e88ed95326b911f198c974cd615ad4