Hook realloc for test harness
This commit implements test_realloc and makes it
possible to adjust the memory size allocated by
test_malloc() or test_calloc().

In this commit, there is still room for improvement.
When we shrink the allocated memory size, we simply
return the original array. This may lead to unused
memory. Therefore, we can improve this in the future.

Reference:
  https://danluu.com/malloc-tutorial/

Change-Id: Ic85b35cfff4dc0b6a46a6e81e6c1ce873625f0ac