Make fmtscan skip inline assembly
Previously, fmtscan would report spelling in inline assembly and block
committing.

It supports the 'asm', '__asm__', and '__asm' keywords. According to the
GCC documentation, the 'asm' keyword is a GNU extension and the
alternative '__asm__' can be used when writing code compiled with
'-ansi' and various '-std' options, and '__asm' is a historical keyword
that the current documentation doesn't mention.

Closes: #292
Change-Id: I9564c47a6ae67278a9b0c42ee0087d597c9c33ab